server/requirements.txt
aioredis==2.0.1
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.73.0
flake8==4.0.1
h11==0.13.0
hiredis==2.0.0
httptools==0.3.0
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
orjson==3.6.6
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==7.0.1
python-dotenv==0.19.2
python-jose==3.3.0
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.1.0
uvicorn==0.17.4
uvloop==0.16.0
watchgod==0.7
websockets==10.1
